Extract TimeDisplaySelectionDialog into its own file
